backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.6.0
sqlalchemy==2.0.34
psycopg[binary,pool]==3.2.1
alembic==1.13.2
python-dotenv==1.0.1
langchain==0.2.16
langgraph==0.2.26
openai==1.51.0
httpx==0.27.2
redis==5.2.0
celery==5.4.0
pytest==8.3.3
pytest-asyncio==0.23.7
httpx==0.27.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-json-logger==2.0.7
prometheus-client==0.20.0
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-instrumentation-httpx==0.48b0
opentelemetry-instrumentation-sqlalchemy==0.48b0
opentelemetry-instrumentation-requests==0.48b0
opentelemetry-instrumentation-openai>=0.20.0
langchain-community==0.2.16
langchain-openai==0.1.21
langchain-core==0.2.36
tiktoken==0.7.0
weaviate-client==4.6.6
lxml==5.3.0
unstructured==0.15.9
pypdf==4.3.1
faiss-cpu==1.8.0
sentence-transformers==3.1.0
pygithub==2.4.0
sentry-sdk[fastapi]==2.14.0
class-variance-authority==0.1.0